refactor(web): decompose CreateAssignmentForm into model + field sections (Tier-2 U15)
